+++ locales/en/common.json
@@ -15,7 +15,7 @@
   "form_IName_text": "Institution Name",
   "form_role_text": "Job Title / Course",
   "form_residence_text": "Province / Country",
-  "form_send_result": "receive results",
+  "form_send_result": "Download results",
   "development": "DEVELOPMENT",
   "most_used_lang": "Java is the most used programming language in Mozambique",
   "most_used_lang_statistics": "Java is the programming language most used by developers in Mozambique. More than 14% of developers use Java, compared to approximately 14% who use Python. On the other hand, Ruby is the programming language least adopted by Mozambicans. ",
