+++ locales/pt/common.json
@@ -15,7 +15,7 @@
   "form_IName_text": "Nome da Instituição",
   "form_role_text": "Título do cargo / Curso",
   "form_residence_text": "Província / País",
-  "form_send_result": "Receber Resultados",
+  "form_send_result": "Baixar os Resultados",
   "development": "DESENVOLVIMENTO",
   "most_used_lang": "Java é a linguagem de programação mais usada em Moçambique",
   "most_used_lang_statistics": "Java é a linguagem de programação mais usada pelos desenvolvedores em Moçambique. Mais de 14% dos desenvolvedores usam Java, contra aproximadamente 14% que usa Python. Por outro lado, Ruby é a linguagem de programação menos aderida pelos moçambicanos.",
