+++ src/app/api/sendEmail/route.ts
@@ -1,14 +1,14 @@
-import sendEmail from "@/utils/sendEmail";
+// import { sendEmail } from "@/utils";
 
-export async function POST(request: Request, response: Response) {
-  const emailDataText = await request.text();
-  const emailData = JSON.parse(emailDataText);
+// export async function POST(request: Request, response: Response) {
+//   const emailDataText = await request.text();
+//   const emailData = JSON.parse(emailDataText);
 
-  try {
-    console.log(emailData);
-    await sendEmail(emailData);
-    return new Response("Email sent successfully", { status: 200 });
-  } catch (error) {
-    return new Response("Error", { status: 500 })
-  }
-};
+//   try {
+//     console.log(emailData);
+//     await sendEmail(emailData);
+//     return new Response("Email sent successfully", { status: 200 });
+//   } catch (error) {
+//     return new Response("Error", { status: 500 });
+//   }
+// }
